split file; ./test passes
