+++ enclone_main/src/lib.rs
@@ -8,6 +8,7 @@ pub mod main_enclone;
 pub mod opt_d_val;
 pub mod sec_mem;
 pub mod setup;
+pub mod start;
 pub mod stop;
 pub mod subset;
 pub mod vars;
