+++ enclone_visual/src/enclone_server.rs
@@ -11,6 +11,7 @@ use enclone_core::combine_group_pics::*;
 use enclone_core::logging::*;
 use enclone_core::parse_bsv;
 use enclone_main::main_enclone::*;
+use enclone_main::start::main_enclone_start;
 use enclone_main::stop::*;
 use enclone_version::*;
 use flate2::write::GzEncoder;
